requirements.txt
fastapi
uvicorn
pillow
opencv-python
numpy
python-multipart
sqlalchemy
alembic
celery
redis
requests
rich
prometheus-fastapi-instrumentator
psycopg2-binary
